+++ appseed/ca/ca4/ca4_file.h
@@ -7,6 +7,7 @@ typedef struct MD5state_st MD5_CTX;
 namespace ca4
 {
 
+
    class CLASS_DECL_ca file :
       public ::radix::object
    {
@@ -40,14 +41,16 @@ namespace ca4
       void is_valid_fileset(const char * pszFile, ::ca::application * papp);
 
       // 'n' (natural) terminated ascii number, example: 245765487n
-      static void write_n_number(ex1::file * pfile, MD5_CTX * pctx, int iNumber);
-      static void read_n_number(ex1::file * pfile, MD5_CTX * pctx, int & iNumber);
+      static void write_n_number(ex1::file * pfile, MD5_CTX * pctx, int64_t iNumber);
+      static void read_n_number(ex1::file * pfile, MD5_CTX * pctx, int64_t & iNumber);
 
       static void write_ex1_string(ex1::file * pfile, MD5_CTX * pctx, string & str);
       static void read_ex1_string(ex1::file * pfile, MD5_CTX * pctx, string & str);
 
 
-
    };
 
+
 } // namespace ca4
+
+
